Get absolute paths to important files and folders appropriate for this computer

Default argument values assume a directory structure that includes Dropbox and is appropriate for the Fellowship project for Dr. Paul Brockway.

get_abs_paths(
  home_path = fs::path_home() %>% as.character(),
  dropbox_path = "Dropbox",
  project_path = file.path(dropbox_path, "Fellowship 1960-2015 PFU database"),
  iea_folder_path = file.path(project_path, "IEA extended energy balance data",
    "IEA 2021 energy balance data"),
  iea_data_path = file.path(iea_folder_path,
    "IEA Extended Energy Balances 2021 (ktoe).csv"),
  country_concordance_path = file.path(project_path, "Mapping",
    "Country_Concordance_Full.xlsx"),
  exemplar_table_path = file.path(project_path, "Database plan", "Exemplar_Table.xlsx"),
  fu_analysis_folder = file.path(project_path, "FU analysis data"),
  ceda_data_folder = file.path(project_path, "Data", "CEDA Data"),
  machine_data_folder = file.path(project_path, "Data", "Machines - Data"),
  reports_source_folders = "reports",
  reports_dest_folder = file.path(project_path, "Reports"),
  drake_cache_folder = file.path(getwd(), "..", "PFU-Database", ".drake")
)

Arguments

home_path
The absolute path to the user's home directory.

dropbox_path
The path to the user's Dropbox directory, relative to home_path.

project_path
The path to the project directory, relative to dropbox_path.

iea_folder_path
The path to the IEA data directory, relative to project_path.

iea_data_path
The path to the IEA data file, relative to iea_folder_path.

country_concordance_path
The path to the country concordance file, relative to project_path.

exemplar_table_path
The path to the exemplar table, relative to project_path.

fu_analysis_folder
The path to the folder containing final-to-useful exergy information, relative to project_path.

ceda_data_folder
The path to the folder containing CEDA (Centre for Environmental Data Analysis) data, relative to project_path.

machine_data_folder
The path to the folder containing machine-specific efficiencies, relative to project_path.

reports_source_folders
A string vector of paths to folders containing report sources.

reports_dest_folder
The path to the folder into which reports will be written, relative to project_path.

drake_cache_folder
The path to the drake cache. Default is ../PFU-Database/.drake. See Details for more information.

Value

A named list containing paths to important directories and files, including:

home_pathThe absolute path to the user's home.

dropbox_pathThe absolute path of the user's Dropbox folder.

project_pathThe absolute path to the project folder.

iea_folder_pathThe absolute path to a folder containing IEA data.

iea_data_pathThe absolute path to the IEA data file for the OECD countries.

country_concordance_pathThe absolute path to the country concordance file.

exemplar_table_pathThe path to the exemplar table.

fu_analysis_folderThe absolute path to the folder containing final-to-useful exergy information.

machine_data_folderThe absolute path to the folder containing machine-specific efficiency information.

ceda_data_folderThe absolute path to the folder containing CEDA environment information.

reports_source_foldersA vector of absolute paths to folders containing source reports.

reports_dest_folderA directory into which completed reports will be written.

drake_cache_folderThe path to the drake cache.

Details

The default argument for home_path gets the value for fs::path_home().

The default value for drake_cache_folder assumes that the PFUSetup folder (this project's folder) is a sibling of the working directory from which the drake workflow is run. If this assumption is not true, users are free to pass a non-default value into this function.

Examples

get_abs_paths()

#> $home_path
#> [1] "/Users/mkh2"
#> 
#> $dropbox_path
#> [1] "/Users/mkh2/Dropbox"
#> 
#> $project_path
#> [1] "/Users/mkh2/Dropbox/Fellowship 1960-2015 PFU database"
#> 
#> $iea_folder_path
#> [1] "/Users/mkh2/Dropbox/Fellowship 1960-2015 PFU database/IEA extended energy balance data/IEA 2021 energy balance data"
#> 
#> $iea_data_path
#> [1] "/Users/mkh2/Dropbox/Fellowship 1960-2015 PFU database/IEA extended energy balance data/IEA 2021 energy balance data/IEA Extended Energy Balances 2021 (ktoe).csv"
#> 
#> $country_concordance_path
#> [1] "/Users/mkh2/Dropbox/Fellowship 1960-2015 PFU database/Mapping/Country_Concordance_Full.xlsx"
#> 
#> $exemplar_table_path
#> [1] "/Users/mkh2/Dropbox/Fellowship 1960-2015 PFU database/Database plan/Exemplar_Table.xlsx"
#> 
#> $fu_analysis_folder
#> [1] "/Users/mkh2/Dropbox/Fellowship 1960-2015 PFU database/FU analysis data"
#> 
#> $machine_data_folder
#> [1] "/Users/mkh2/Dropbox/Fellowship 1960-2015 PFU database/Data/Machines - Data"
#> 
#> $ceda_data_folder
#> [1] "/Users/mkh2/Dropbox/Fellowship 1960-2015 PFU database/Data/CEDA Data"
#> 
#> $reports_source_folders
#> [1] "reports"
#> 
#> $reports_dest_folder
#> [1] "/Users/mkh2/Dropbox/Fellowship 1960-2015 PFU database/Reports"
#> 
#> $drake_cache_folder
#> [1] "/Users/mkh2/github/PFUSetup/docs/reference/../PFU-Database/.drake"
#>